go_to_frame: back up problem solved
AMCL parameters `update_min_a` (update on angular motion) and
`update_min_d` (update on linear motion) have to be positive. In
simulator it caused the "forward" movement in Rviz2 while the robot was
moving backward, and in the lab it caused negative eigenvalues to be
computed for position and general block of AMCL and `/initialpose`.